feat: adder encoding; solution phase-saving for LSU; predict number of aux. clauses for GTE

+++ MaxSAT.cc
@@ -350,7 +350,8 @@ void MaxSAT::printBound(int64_t bound)
   if(!print) return;
 
   // print bound only, if its below the hard weight
-  if( bound < (int64_t)maxsat_formula->getHardWeight() ) printf("o %" PRId64 "\n", bound);
+  // FIXME: possible issue for PB instances when bound is negative; in MaxSAT bound is always positive
+  if( bound < maxsat_formula->getHardWeight() ) printf("o %" PRId64 "\n", bound);
 }
 
 // Prints the best satisfying model. Assumes that 'model' is not empty.
